[NS-72] Reorganize `aveobench` code
- Move column vector related code inside `aveobench` to
`io.sparkcyclone.data.vector` package
