+++ aveobench/src/main/scala/io/sparkcyclone/benchmarks/InputSamples.scala
@@ -1,5 +1,6 @@
 package io.sparkcyclone.benchmarks
 
+import io.sparkcyclone.data.VeColVectorSource
 import io.sparkcyclone.data.vector._
 import io.sparkcyclone.data.conversion.SeqOptTConversions._
 import scala.reflect.ClassTag
@@ -42,10 +43,10 @@ object InputSamples {
 
   def bpcv(typ: String, ncolumns: Int, size: Int)(implicit source: VeColVectorSource): Seq[BytePointerColVector] = {
     typ match {
-      case "Short"  => 0.until(ncolumns).map(_ => InputSamples.seqOpt[Short](size).toBytePointerColVector("_"))
-      case "Double" => 0.until(ncolumns).map(_ => InputSamples.seqOpt[Double](size).toBytePointerColVector("_"))
-      case "String" => 0.until(ncolumns).map(_ => InputSamples.seqOpt[String](size).toBytePointerColVector("_"))
-      case _        => 0.until(ncolumns).map(_ => InputSamples.seqOpt[Int](size).toBytePointerColVector("_"))
+      case "Short"  => 0.until(ncolumns).map(_ => seqOpt[Short](size).toBytePointerColVector("_"))
+      case "Double" => 0.until(ncolumns).map(_ => seqOpt[Double](size).toBytePointerColVector("_"))
+      case "String" => 0.until(ncolumns).map(_ => seqOpt[String](size).toBytePointerColVector("_"))
+      case _        => 0.until(ncolumns).map(_ => seqOpt[Int](size).toBytePointerColVector("_"))
     }
   }
 }
