+++ aveobench/src/main/scala/io/sparkcyclone/data/vector/CompressedVeColBatch.scala
@@ -1,9 +1,9 @@
-package io.sparkcyclone.colvector
+package io.sparkcyclone.data.vector
 
 import io.sparkcyclone.data.vector._
 import io.sparkcyclone.vectorengine.VeProcess
 
-final case class CompressedVeColBatch private[colvector] (columns: Seq[UnitColVector],
+final case class CompressedVeColBatch private[vector] (columns: Seq[UnitColVector],
                                                           struct: Long,
                                                           buffer: Long) {
   def free()(implicit process: VeProcess): Unit = {
